Remove URI::URL.  Support git scheme.
URI::URL::strict() has been replaced with a check to see if the URI object
is of type URI::_foreign which indicates a scheme URI does not understand.
This is a violation of encapsulation.  It's been very stable for years
now and I'm going with it for the time being.

Best solution would be to have our own list of schemes, but I don't
have an internet connection right now to handle that.

URI::URL would always use the canonical version.  URI stringifies to the
exact URL as written.  I think this is better, less loss of information.
You can always ask for the canonical version.  This might cause a
backwards compat problem for some users.

Added the git scheme as our first non-URI.pm supported scheme.